# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "BudgetBuddy API",
    "description": "API for analyzing and reporting user and group expenses.",
    "version": "1.1"
  },
  "servers": [
    {
      "url": "http://localhost:3000"
    }
  ],
...
  "paths": {
...
    "/api/expenses": {
      "post": {
        "tags": ["Expense Tracking Service"],
        "summary": "Create a new expense",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Expense"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Expense created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/expenses/group/{groupId}": {
      "get": {
        "tags": ["Expense Tracking Service"],
        "summary": "Get expenses for a specific group",
        "parameters": [
          {
            "name": "groupId",
            "in": "path",
            "description": "Unique identifier of the group",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of expenses for the specified group",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Expense"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Group not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/expenses/user": {
      "get": {
        "tags": ["Expense Tracking Service"],
        "summary": "Get expenses for the authenticated user",
        "responses": {
          "200": {
            "description": "A list of expenses for the authenticated user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Expense"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/api/expenses/{expenseId}": {
      "get": {
        "tags": ["Expense Tracking Service"],
        "summary": "Get a specific expense by ID",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "description": "Unique identifier of the expense",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed information about the specified expense",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpenseWithDetails"
                }
              }
            }
          },
          "404": {
            "description": "Expense not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "put": {
        "tags": ["Expense Tracking Service"],
        "summary": "Update an existing expense",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "description": "Unique identifier of the expense to be updated",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Expense"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Expense updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "400": {
            "description": "Missing required fields"
          },
          "404": {
            "description": "Expense not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      },
      "delete": {
        "tags": ["Expense Tracking Service"],
        "summary": "Delete a specific expense",
        "parameters": [
          {
            "name": "expenseId",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Expense deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Expense"
                }
              }
            }
          },
          "404": {
            "description": "Expense not found"
          },
          "500": {
            "description": "Internal server error"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Expense": {
        "required": ["amount", "budgetId", "date", "userId"],
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "float"
          },
          "budgetId": {
            "type": "string",
            "format": "objectId",
            "description": "The unique identifier of the Budget, in MongoDB ObjectId format."
          },
          "categoryId": {
            "type": "string",
            "format": "objectId",
            "description": "The unique identifier of the ExpenseCategory, in MongoDB ObjectId format."
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "groupId": {
            "type": "string",
            "format": "objectId",
            "description": "The unique identifier of the Group, in MongoDB ObjectId format."
          },
          "_id": {
            "type": "string",
            "format": "objectId",
            "description": "The unique identifier of the Expense, in MongoDB ObjectId format."
          },
          "userId": {
            "type": "string",
            "format": "objectId",
            "description": "The unique identifier of the User, in MongoDB ObjectId format."
          }
        }
      },
...
      "ExpenseWithDetails": {
        "type": "object",
        "properties": {
          "amount": {
            "type": "number",
            "format": "float"
          },
          "budgetId": {
            "type": "string",
            "format": "objectId",
            "description": "The unique identifier of the Budget, in MongoDB ObjectId format."
          },
          "category": {
            "$ref": "#/components/schemas/ExpenseWithDetails_category"
          },
          "date": {
            "type": "string",
            "format": "date"
          },
          "description": {
            "type": "string"
          },
          "group": {
            "$ref": "#/components/schemas/ExpenseWithDetails_group"
          },
          "_id": {
            "type": "string",
            "format": "objectId",
            "description": "The unique identifier of the Expense, in MongoDB ObjectId format."
          },
          "user": {
            "$ref": "#/components/schemas/GroupWithDetails_members"
          }
        }
      },
      "ExpenseWithDetails_category": {
        "type": "object",
        "properties": {
          "categoryName": {
            "type": "string"
          }
        }
      },
      "ExpenseWithDetails_group": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "objectId",
            "description": "The unique identifier of the Group, in MongoDB ObjectId format."
          },
          "groupName": {
            "type": "string"
          }
        }
      },
...
      "GroupWithDetails_members": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string",
            "format": "objectId",
            "description": "The unique identifier of the User, in MongoDB ObjectId format."
          },
          "email": {
            "type": "string",
            "format": "email"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}